server/requirements.txt
# requirements_server.txt
fastapi==0.117.1
uvicorn==0.37.0
websockets==15.0.1
gunicorn

# Core dependencies for the packages above
anyio==4.11.0
click==8.1.8
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
watchfiles==1.1.0
PyYAML==6.0.2

sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-auth